refactor: update wallet hook imports to new module structure for better organization
